+++ SHSpecial_C/Copy/DailyStructs.py
@@ -5,6 +5,6 @@ from ctypes import c_uint64
 
 class RateValueItem(Structure):
 	_fields_ = [("isDayActive",c_bool),
-		("backrange",c_int64),
-		("forrange",c_int64),
+		("backrange",c_int32),
+		("forrange",c_int32),
 		("filler",c_uint64*4)]
